+++ ci/cpu/build.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 # Copyright (c) 2018, NVIDIA CORPORATION.
 ######################################
-# cuDF CPU conda build script for CI #
+# cuXfilter CPU conda build script for CI #
 ######################################
 set -e
 
@@ -49,8 +49,11 @@ conda list
 conda config --set ssl_verify False
 
 ################################################################################
-# BUILD - Conda package builds (conda deps: libcudf <- libcudf_cffi <- cudf)
+# BUILD - Conda package build
 ################################################################################
+
+#logger "Cloning cuDataShader..."
+#git clone git@github.com:rapidsai/cuDataShader.git
 
 logger "Build conda pkg for cuxfilter..."
 source ci/cpu/cuxfilter/build_cuxfilter.sh
@@ -61,3 +64,4 @@ source ci/cpu/cuxfilter/build_cuxfilter.sh
 
 logger "Upload conda pkgs..."
 source ci/cpu/upload_anaconda.sh
+