+++ ci/gpu/build.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 # Copyright (c) 2018, NVIDIA CORPORATION.
-#########################################
-# cuDF GPU build and test script for CI #
-#########################################
+##############################################
+# cuXfilter GPU build and test script for CI #
+##############################################
 set -e
 NUMARGS=$#
 ARGS=$*
@@ -44,7 +44,8 @@ logger "Activate conda env..."
 source activate gdf
 conda install "cudf=$MINOR_VERSION.*" "cudatoolkit=$CUDA_REL" \
               "numpy>=1.16" "cupy>=6.0.0" "pandas>=0.24.2,<0.25" "panel=0.6.*" \
-              "bokeh>=1.2.*" "geopandas>=0.6.*" "pyproj=1.9.*" "pytest"
+              "bokeh>=1.2.*" "geopandas>=0.6.*" "pytest" "pyppeteer" \
+              "pyproj>=2.4.*"
 
 # Install the master version of cudatashader
 logger "pip install git+https://github.com/rapidsai/cuDataShader.git --upgrade --no-deps"